+++ docs/change-log.md
@@ -15,6 +15,7 @@ outline: deep
 * zodText() and zodMediumText() have become deprecated aliases for zodString().
 * Implemented: make default a required field in a prop definition #127
 * Fixed /api/v1/user/:id/s3files
+* Added ManageButton component.
 
 ## 2024-04-25
 * Fixed a bug where ving record fields of type 'int' were not being initialized properly.